Move collect stuff to new file
